pyproject.toml
[project]
name = "bioideas"
version = "0.1.0"
description = "Extract and evaluate biotech startup ideas from podcast transcripts"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.9.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.4.0",
    "hdbscan>=0.8.33",
    "tqdm>=4.66.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "streamlit>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
bioideas = "bioideas.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
